# Conflicts:
#	web/src/main/java/ua/com/vertex/controllers/HomePageController.java
#	web/src/main/java/ua/com/vertex/dao/UserDao.java
#	web/src/main/webapp/WEB-INF/web.xml
#	web/src/main/webapp/css/bootstrap-theme.min.css
#	web/src/main/webapp/css/bootstrap.min.css
#	web/src/main/webapp/home.jsp

+++ web/src/main/java/ua/com/vertex/dao/UserDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface UserDao {
 
     @SuppressWarnings("unused")
-    User getUser(int id);
+    User getUser(long id);
 
     @SuppressWarnings("unused")
     void deleteUser(long id);
